fix: Chrome TextTrack bug w/ cue refresh
